reorganize math
