+++ Implementations/11 - Math (4)/Operators/modOp.cpp
@@ -2,9 +2,18 @@
 * Description: Basic operations with modular arithmetic
 */
 
-namespace ModOp {
+namespace modOp {
     ll po (ll b, ll p) { return !p?1:po(b*b%MOD,p/2)*(p&1?b:1)%MOD; }
     ll inv (ll b) { return po(b,MOD-2); }
+    
+    ll inv(ll a, ll b) { // 0 < a < b, gcd(a,b) = 1
+        a %= b;
+        if (a <= 1) return a;
+        ll i = inv(b%a,a);
+        ll tmp = -((b/a)*i+((b%a)*i)/a) % b;
+        if (tmp < 0) tmp += b;
+        return tmp;
+    }
     
     int ad(int a, int b) { return (a+b)%MOD; }
     int sub(int a, int b) { return (a-b+MOD)%MOD; }
