+++ Math/Math Contest Scoring.gs
